Collapse copy-sync/ and move-sync/ into the async folders
