--- lib/copy-sync/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-module.exports = {
-  copySync: require('./copy-sync')
-}
