+++ lib/copy/__tests__/copy-sync-broken-symlink.test.js
