+++ lib/copy/__tests__/copy-sync-case-insensitive-paths.test.js
