+++ lib/copy/__tests__/copy-sync-dir.test.js
