+++ lib/copy/__tests__/copy-sync-file.test.js
