+++ lib/copy/__tests__/copy-sync-prevent-copying-identical.test.js
