+++ lib/copy/__tests__/copy-sync-prevent-copying-into-itself.test.js
