+++ lib/copy/__tests__/copy-sync-readonly-dir.test.js
