+++ lib/copy/__tests__/copy-sync-symlink.test.js
