+++ lib/copy/copy-sync.js
