+++ lib/copy/index.js
@@ -2,5 +2,6 @@
 
 const u = require('universalify').fromCallback
 module.exports = {
-  copy: u(require('./copy'))
+  copy: u(require('./copy')),
+  copySync: require('./copy-sync')
 }
