--- lib/move-sync/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-module.exports = {
-  moveSync: require('./move-sync')
-}
