+++ lib/move/__tests__/move-sync-case-insensitive-paths.test.js
