+++ lib/move/__tests__/move-sync-prevent-moving-identical.test.js
