+++ lib/move/__tests__/move-sync-prevent-moving-into-itself.test.js
