+++ lib/move/__tests__/move-sync.test.js
