+++ lib/move/index.js
@@ -2,5 +2,6 @@
 
 const u = require('universalify').fromCallback
 module.exports = {
-  move: u(require('./move'))
+  move: u(require('./move')),
+  moveSync: require('./move-sync')
 }
