+++ lib/move/move-sync.js
@@ -2,7 +2,7 @@
 
 const fs = require('graceful-fs')
 const path = require('path')
-const copySync = require('../copy-sync').copySync
+const copySync = require('../copy').copySync
 const removeSync = require('../remove').removeSync
 const mkdirpSync = require('../mkdirs').mkdirpSync
 const stat = require('../util/stat')
